Update README for instance state changes (#479)
NTH's SQS processor does automatically support instance state changes for EC2 instances. If a node is Terminated, either by a user or by an AWS scheduled event, we still have a small window in which to begin the drain. 

The documentation as it stands doesn't set up the EventBridge rule to enable this. I've added it to the basic setup guide here.